+++ src/oxylabs/proxy/proxy.py
@@ -1,10 +1,8 @@
 from typing import Optional
 from urllib.parse import quote, urlparse
-
 import requests
-
 from src.oxylabs.utils.defaults import NON_UNIVERSAL_DOMAINS, PROXY_BASE_URL, PROXY_PORT
-
+from src.oxylabs.utils.utils import prepare_config
 
 class Proxy:
     def __init__(self, username: str, password: str) -> None:
@@ -35,24 +33,26 @@ class Proxy:
         """
         return f"http://{self._username}:{self._password}@{PROXY_BASE_URL}:{PROXY_PORT}"
 
-    def get(self, url: str) -> Optional[requests.Response]:
+    def get(self, url: str, request_timeout: Optional[int] = None) -> Optional[requests.Response]:
         """
         Sends a GET request to the specified URL using the session object.
 
         Args:
             url (str): The URL to send the GET request to.
+            request_timeout (Optional[int]): The request timeout in seconds. Defaults to None (no timeout).
 
         Returns:
             Optional[requests.Response]: The response object returned by the GET request, or None if an error occurred.
-
-        Raises:
-            requests.exceptions.RequestException: If the GET request encounters an error.
         """
         try:
+            config = prepare_config(request_timeout=request_timeout)
             self._url_to_scrape = url
-            response = self._session.get(url)
+            response = self._session.get(url, timeout=config["request_timeout"])
             response.raise_for_status()
             return response
+        except requests.exceptions.Timeout:
+            print(f"Timeout error. The request to {url} has timed out after {request_timeout} seconds.")
+            return None
         except requests.exceptions.RequestException as e:
             print(f"Request failed: {e}")
             return None