+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+
+- Fixed function naming.
+- Added request timeout in proxy.
+
 ## 1.0.1
 
 - Fixed issue with uploaded package.
@@ -7,7 +12,7 @@
 ## 1.0.0
 
 - Initial release of Oxylabs SDK.
-- Web scraping categories:
+- Scraper APIs:
   - SERP
   - Ecommerce
 - Integration methods: